+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 * Support single level module imports of names too, such as using o
   ``from django.utils import crypto`` with ``crypto.get_random_string()``.
 
+* Add fixer for Django 3.1 ``NullBooleanField`` deprecation.
+
 1.1.0 (2021-08-28)
 ------------------